Enable Trabajo Actual checkbox in edit modified
if the checkbox for 'is currently working' is checked, this enables
the textbox asking for the name of the Candidato's current employer,
if not, then shows it as disabled.

this fixes the issue when editing a Candidato that is saved with
the checkbox checked, which will show it as disabled at the start,
which it shouldn't, but then enable the textbox when the checkbox
is not checked. It shouldn't enable the textbox if the Candidato
is not currently working, and vice versa.